move some types definitions to a common package to be reused by others circuits
